pyproject.toml
[tool.poetry]
name = "copyblock"
version = "0.1.0"
description = ""
authors = ["MarioGom"]
license = "CC0"

[tool.poetry.dependencies]
python = "^3.8"
pywikibot = "^6.4.0"
python-dateutil = "^2.8.2"
mwparserfromhell = "^0.6.2"
click = "^8.0.1"
requests-oauthlib = "^1.3.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
mypy = "^0.910"
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
